Convert BubbleChart to tsx, improve ExampleView responsiveness
