+++ www/src/styles/container.scss
@@ -45,8 +45,6 @@ footer {
 .page-guide,
 .page-api,
 .page-examples {
-  max-width: 1200px;
-  margin: 0 auto;
   padding: 0 20px;
 }
