+++ www/src/views/ExampleView.scss
@@ -13,7 +13,7 @@
       border: 1px solid #f5f5f5;
     }
 
-    @media screen and (max-width: 1300px) {
+    @media screen and (max-width: 1600px) {
       .example-inner-wrapper {
         flex-direction: column;
